Small tweaks to board
The MOSFET for the LEDs was a TO-247 which did not match the part I had selected. Changed it to a TO-220 to match the part. Also added a pull-down resistor to the gate to ensure it turns off when there is no longer a voltage on the pin. The PIC chip will ground it but it makes sense to have the extra part in case the board is used without it.

Signed-off-by: Nathan Hjelm <hjelmn@me.com>