Small tweaks
The original schematic I used to design the board erroneously used pin 4 (GPIO3) as an output pin. This will not work with a PIC2F675 as that pin is input-only. Moved the output to GPIO4 to correct the issue and adjusted some other components based on observations.

To fix the boards I have I put a jumper between GPIO3 and GPIO4 and disabled external MCLR to avoid the output causing a reset of the chip.

Also, can use a 200 ohm resistor in place of the potentiometer (dropped to 1k). 10k might be a bit too high but could work as well. Will do more calculations to see what is best since the 10k can only go down to 5.3k. At 200 ohm it takes about 8 clock cycles for the GPIO pin to register high at the appropriate light level.

Signed-off-by: Nathan Hjelm <hjelmn@me.com>