Update with a 4:0 LEDs:resistors ratio
After closer inspection it seems the original sign was using a 4:1 ratio given there are 3 resistors for each eye/eyebrow (12 LEDS), 3 for the right arm (12 LEDs), etc. The LEDs I ordered also appear to be 2.6-2.9v forward voltage drop (not the 3-3.2 advetised) so it is better to do 4 LEDs to one 64 ohm resistor to get ~ 10mA of current. These LEDs are super bright so running close to 20mA would be a mistake.

Signed-off-by: Nathan Hjelm <hjelmn@me.com>